Add Publish subcommand, configurable authorization header
This will be necessary for the changes coming to the QPM backend soon. Also added it as a configuration option, but it cannot be saved via CLI at the moment. This may change, but if it does, it should change such that the local config NEVER holds it (likewise, the local config should NEVER hold the GH credentials)